1.1 Add Cross Origin Configuration 1.2 Add Exception Handler Middleware 1.3 Tweak Cart and Order Items 1.4 Add Get Current User To Auth Controller 1.5 Add Seeder Base and Product Seeder

+++ Shopping.Application/Services/Concrete/OrderItemService.cs
@@ -3,6 +3,7 @@ using Shopping.Application.Http.Exceptions;
 using Shopping.Application.Http.Exceptions.Order;
 using Shopping.Application.Resources.Order.OrderItem;
 using Shopping.Application.Services.Abstract;
+using Shopping.Application.Validations.Order;
 using Shopping.Domain.Entities;
 using Shopping.Domain.Shared.Enums;
 using Shopping.EntityFrameworkCore.Repositories.Abstract;
@@ -27,7 +28,7 @@ public class OrderItemService : IOrderItemService
 
     public async Task<ICollection<OrderItemResponse>> GetOrderItems(Guid orderId, Guid userId)
     {
-        Order? order = await _orderRepository.GetOrderWithOrderItems(orderId);
+        Order? order = await _orderRepository.GetOrderWithOrderItemsAsync(orderId);
         if (order == null) throw new NotFoundException();
         if (order.UserId != userId) throw new UnauthorizedException();
 
@@ -36,7 +37,11 @@ public class OrderItemService : IOrderItemService
 
     public async Task<OrderItemResponse> IncreaseQuantity(Guid orderId, Guid productId, Guid userId)
     {
-        var (order, orderItem) = await ValidateOrder(orderId, productId, userId);
+        var (order, orderItem) = await ValidateOrderCommon(orderId, productId, userId);
+
+        int orderItemsTotalQuantity = await _orderItemRepository.SumOrderItemTotalQuantityByOrderIdAsync(orderId);
+        if (orderItemsTotalQuantity == OrderValidations.OrderItems.MinCount) throw new OrderItemsTotalQuantityLimitExceededException();
+
         orderItem.Quantity++;
         order.TotalPrice += orderItem.Price;
         await _unitOfWork.SaveChangesAsync();
@@ -46,10 +51,12 @@ public class OrderItemService : IOrderItemService
 
     public async Task<OrderItemResponse> DecreaseQuantity(Guid orderId, Guid productId, Guid userId)
     {
-        var (order, orderItem) = await ValidateOrder(orderId, productId, userId);
+        var (order, orderItem) = await ValidateOrderCommon(orderId, productId, userId);
         if (orderItem.Quantity == 1) throw new OrderItemMinQuantityException();
+
         orderItem.Quantity--;
         order.TotalPrice -= orderItem.Price;
+
         await _unitOfWork.SaveChangesAsync();
 
         return _mapper.Map<OrderItem, OrderItemResponse>(orderItem);
@@ -57,20 +64,26 @@ public class OrderItemService : IOrderItemService
 
     public async Task Remove(Guid orderId, Guid productId, Guid userId)
     {
-        var (order, orderItem) = await ValidateOrder(orderId, productId, userId);
+        var (order, orderItem) = await ValidateOrderCommon(orderId, productId, userId);
+
+        var orderItemCount = await _orderItemRepository.CountOrderItemByOrderIdAsync(orderId);
+        if (orderItemCount == 1) throw new MinimumOrderItemsNotSatisfiedException();
+
         order.TotalPrice -= orderItem.Price * orderItem.Quantity;
         _orderItemRepository.Remove(orderItem);
+        
         await _unitOfWork.SaveChangesAsync();
     }
 
-    private async Task<Tuple<Order, OrderItem>> ValidateOrder(Guid orderId, Guid productId, Guid userId)
+    private async Task<Tuple<Order, OrderItem>> ValidateOrderCommon(Guid orderId, Guid productId, Guid userId)
     {
         Order? order = await _orderRepository.GetByIdAsync(orderId);
+
         if (order == null) throw new NotFoundException();
         if (order.UserId != userId) throw new UnauthorizedException();
         if (order.Status == OrderStatus.Submitted) throw new SubmittedOrderCanNotEditException();
 
-        OrderItem? orderItem = await _orderRepository.GetOrderItemById(productId, userId);
+        OrderItem? orderItem = await _orderRepository.GetOrderItemByIdAsync(orderId, productId);
         if (orderItem == null) throw new NotFoundException();
 
         return Tuple.Create(order, orderItem);
